Standard / ISO19115-3 / Processing / Language refactor
Add a process that can manage language configuration for a record:
* set default language (preserving translations from an additional language if exist)
* reorder additional languages
* remove all additional languages (ie. create a monolingual record)

This can be useful when importing records from a multilingual catalogue and wants to preserve only one language on reception (eg. when third party apps using the catalogue does not support multilingual records).

eg.

* `ISO19139-to-ISO19115-3-2018-with-languages-refactor?defaultLanguage=fre&others=none`
* `ISO19139-to-ISO19115-3-2018-with-languages-refactor?defaultLanguage=fre&others=eng`
* `ISO19139-to-ISO19115-3-2018-with-languages-refactor?defaultLanguage=fre&others=ger,eng`

Also cleanup an old unused test folder and create tests for the language-refactor process.